+++ runtime/CHANGES.md
@@ -1,4 +1,5 @@
 #### Version: 0.3.2 (LibGDX 1.9.2, Artemis 1.3.1)
+- **API Addition**: constructor `VisSpriter (Loader<Sprite> loader, Data data, float scale, int entityIndex)`
 
 #### Version: 0.3.1 (LibGDX 1.9.2, Artemis 1.3.1)
 - Updated to LibGDX 1.9.2 and Artemis 1.3.1